requirements.txt
amqp==5.2.0
argh==0.31.3
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.8.0
celery==5.4.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.8
django-appconf==1.0.6
django-autoslug==1.9.9
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-celery-results==2.5.1
django-countries==7.6.1
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
flake8==7.1.1
flower==2.0.1
html5lib==1.1
humanize==4.10.0
idna==3.10
isort==5.13.2
kombu==5.4.2
lxml==5.3.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.45
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.1
typing-extensions==4.14.0
tzdata==2024.2
urllib3==2.5.0
vine==5.1.0
watchdog==5.0.2
wcwidth==0.2.13
webencodings==0.5.1
selenium==4.35.0
pytest==8.3.2
pytest-django==4.9.0
pytest-cov==5.0.0
